requirements.txt
fastapi
uvicorn
gunicorn
ultralytics
pillow
numpy
torch
torchvision
requests
python-multipart